app with parser schedule
